+++ scripts/codex/constants.js
@@ -49,6 +49,7 @@ export const STATUS_ARMOR_BROKEN = "qol-broken-armor";
 export const FLAG_CHARACTER_CREATION = "greybeardCharacterCreation";
 export const FLAG_CREW_ROSTER = "crewRoster";
 export const FLAG_TRAINING_SKILL = "trainingSkill";
+export const FLAG_CYBERWARE = "cyberware";
 
 export const SETTING_ENABLE_CHARACTER_CREATOR = "enableCharacterCreator";
 export const SETTING_THEME_COLOR = "themeColor";